Generalize the Capabilities class and move it to exported package
Once in a while there is a demand to find providers for a requirement in
a way that is not covered by the current wiring state, e.g. when
traversing a repository or some other state of Resources. As indexing is
hard to get done without access to some framework internals only using
public API is not trivial. Equinox already contains a Capabilities class
that supports such indexing using internal of the framework impl.

This makes the Capabilities class generic, so it can be used with any
way of obtaining capabilities from a resource and move it to a public
package.