Make var-args overload of NLS.bind() generic
This avoids a compiler warning if an explicit array of a sub-type of
Object[] is passed as last argument:
NLS.bind("message: {0} {1}", new String[] { "Hello", "World" })

The byte-code generated for the NLS.bind() method and for all its
callers is not changed by this adjustment. It just avoids what looks
like a source-compatibility issues with the previous introduction of
var-args in NLS.bind(). Although the example above works the same in all
three states.

Follow-up on:
- https://github.com/eclipse-equinox/equinox/pull/936

Also touch calling projects that now get slight different byte-code
generated because the var-args array generated by the compiler now can
have a more specific type than Object[], common for all passed
arguments.