[Build] Remove check for launcher binary rebuild commit
Currently the build is aborted if the branch tip is at a commit, which
was committed with the email of the 'Eclipse Equinox Bot'. This was done
with the intention to prevent (endless) rebuilds after a commit for a
launcher binary rebuild was pushed by the bot. But because that bot is
also used for other kind of commits, this test leads to false positives.

An in-depth analysis showed that the pipeline does in fact not get into
an endless native rebuild loop in case that check is removed. Only one
new build is triggered, when a commit for a launcher binary rebuild is
pushed (as usual for a push), which does not rebuild the native binaries
again because the latest launcher-binaries-build tag points to the then
HEAD commit.
So this reworks the pipeline to just remove that check.