Do not print optional unresolved requirements in report
Currently ModuleResolutionReport prints all unresolved requirements what
in case of optional requirements can bloat the report considerably and
is often confusing to read as optional requirements can not lead to an
unresolved state.

This now filters out all items from the report if the requirement is
optional and adding a system property for the case someone wants to
still enable the old behavior.