Throw exception when last candidate is removed and mandatory requirement
Currently it can happen that the resolver is removing a candidate but it
is the last possible one to select. If the requirement is not optional
this lead guaranteed to an unresolvable state and must be handled
already on the call site.

This adds throwing a runtime exception for this case to make sure such
illegal modifications do not goes unnoticed and we can fix such places.