Extract inner classes from ResolverImpl
Currently the ResolverImpl is really hard to understand and maintain as
it has > 2600 LOC. One reason is that it contains many static inner
classes and ResolveSession is even already package protected to be used
in other places.

This now extract all inner static (final) classes into own types, this
reduces the size of the ResolverImpl to 1800+ LOC and makes it easier to
enhance in the future.