Throw Exception from getInitialCandidates directly
getInitialCandidates currently set the error in the session instead of
throwing an exception, but the only caller of that method just checks if
an error is set and then throws it anyways.

This now just throw the error right away from the method to make the
exception trace more meaningful.