Pull out Candidates#checkSubstitutes from ResolverImpl#checkConsistency
Currently checkConsistency calls Candidates#checkSubstitutes first and
if any error is found return it immediately. This is not only confusing
and harder to debug, but also a missing requirement found in this
process does not mean the solution is inconsistent it is only not fully
resolved.

Because of this and to make further improvements possible here this now
extracts the Candidates#checkSubstitutes from the
ResolverImpl#checkConsistency.