Log and assert actually processed permutations
Currently we log the actual number of permutation removed from the stack
and also assert the counts for the individual permutation types added.
Now as we have the Backlog actually filtering out candidates that are
likely not good solutions the actual number of processed permutations
becomes much more interesting.

This now adds a new logger method for that case and uses it in the tests
to assert that no more than a certain number of permutations are
actually processed in the resolver loop.