Improve substitution package handling and filter uses locally
Currently substitution packages handling can lead to a situation where
the number of performed permutation increase considerably and the
finding of a solution takes very long. In some cases it even times out
and results in a failed resolve state.

This now is improved by the following changes:

When a permutation is removed from the stack, every requirement is
checked if it is a substitution package and then resolve it to either
external or internal case first:
- if more candidates exits for this requirement a permutation is added
to account for this alternative solution
- now all other providers except the current one are dropped to make
this a single provider as the choice can not be reverted and therefore
other permutations in this path will be invalid
- then it must be checked if the package resolves to an external
provider in which case we need to drop the exported package capabilities
of this bundle for the given package
- after that all requirements are checked if they have only a single
provider (what can happen independently or as part of the resolving to
either external or internal) the use constraints are checked if they are
in conflict with a currently selected provider and dropping such invalid
choices
- finally after that we check the package space consistency (that is a
far more expensive check) what probably will result in more permutations
created and checked until we found the best result to use