Use empty candidates instead of null when removing from the map
Currently in certain situations we remove the requirement form the map
completely when no providers are left, but this results in the
FaultyResourcesReport to not working properly and assume a permutation
is valid / better than it is.

This now replaces the entry in the map with an empty instance instead of
completely removing it.