Unify access to the selector for a requirement
Currently there are many places that access the map to get the selector,
most places check for null but one also checks for empty and another is
only checking for empty.

To unify this, there is now a method getSelector that is used in all
places to unify the behavior to always return null when empty.