Use faulty resources as a first discriminator
Currently we extract the failed requirement of a use-constrain violation
but this actually results in the whole resource to be unresolved. This
leads to the situation that a solution that has one optional requirement
not satisfied is compared equal to one that is failing on an optional
requirement with uses-violation what is wrong.

This now store the map directly and uses it as a first selection
criteria to decide one solution is better than the other