Init system module and assert all modules are resolved
When using a module database as a testcase this now makes the actual
resolving a bit more like it happens on runtime that is the system
module is init (and resolved) first and then the other bundles are
resolved. Beside that we also assert that all modules are resolved in
the end to make sure we not just missing them in the report.