Do not depend on PlatformAdmin for computing the container timestamp
Currently equinox.registry uses the old compatibility service
PlatformAdmin to gather a timestamp that can be used to detect if
something changed. In platform we already migrated to an approach that
uses the last modified of the modules instead.

This now applies the same here to get one less reference to the
PlatformAdmin service.