Add a Candidates#toString method to make debugging easier
Currently debugging problems in the resolver involves a comparing
different alternatives solutions. As the overall state could be quite
big inspecting the state of the object itself can be quite challenging.

This now adds a new Candidates#toString method that allows for easy
print out of the state of a Candidates object enriching it with some
extracted data like if a requirement is optional or a substitution
package and what a resource currently provides to others.