Propagate the cause of the failure from loadClass
If currently the bundle can not be resolved the method Bundle#loadClass
fails with a quite generic exception that it hard to understand as it
does not contain the name of the class to load nor the actual root cause
that the bundle is unable to resolve.

This now move the actual logging out of the place of fetching and
resolving method to the loadCLass method that allows to also throw a
proper exception message here.