Add a new method to create a copy of BundleDescription with new ID
When creating a BundleDescription that is a copy of an exiting one it
still inherits its ID. This is problematic when one created a new state
and want to combine bundles from different state into this e.g. to
perform a new resolve operation.

This now adds a new method that allows to copy the bundle description
together with assign a new ID to it.