Add a new thin abstraction layer over the creation of resolver instances
Currently Equinox directly reference the implementation of the (felix)
resolver even though it never uses any specific API (except the logger),
this makes it currently hard to allow other implementations especially
as the is an unfortunate relation between the startup of launches and
tests, so if another  implementation is used and might be buggy the
framework possibly not start up and tests can not be executed or
debugged.

This now adds a thin abstraction layer over the creation of resolver
instances that basically distinguish the three cases we have where an
instance is created allowing to later e.g. return dynamically based on
the context (e.g. testrun) a different resolver instance.