Temporarily restore EquinoxBundle#getModuleClassLoader(boolean)
The private method EquinoxBundle#getModuleClassLoader(boolean) was
recently replaced by a method with a different signature. Some relevant
consumers (at least Jetty OSGi's DefaultBundleClassLoaderHelper)
reflectively access the replaced method. To avoid that compatibility of
such consumers breaks, this change temporarily reintroduces the replaced
method. It prints an according warning about the upcoming removal to
stderr. It is supposed to be removed once the known consumers are
adapted.