Migrate RegionPerformanceTests to JUnit 5
RegionPerformanceTests currently relies on a TestPerformanceRunner
method that is specific to JUnit 3 (using the TestCase class). In order
to make TestPerformanceRunner independent from JUnit 3, this change
replaces the usage of that method with a different method while
preserving the behavior. It also completely migrates the performance
test class to JUnit 5 and cleans up unused method parameters.